Changes in view functions, templates and django paper search, minor change in forms
Lots of changes in templates:
1. display: 1 result card per row
2. highlighting of annotation in the cited paper and author search results.

Also, in views and django paper search, there's a sorting by published date. There's some rearrangement of the code for normalizing, and of course, sending the highlighted part to the template in the view

In forms, add max_values to all the num rows fields.